backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
groq==0.4.2
langchain==0.1.0
langgraph==0.0.20
langchain-groq==0.0.1
langchain-core==0.1.24
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
